Move image masking utility to tests
